moved checkConnectivity to ClusterCli and modified code to not expose NetworkServer in WaltzServer.
