+++ waltz-tools/src/main/java/com/wepay/waltz/tools/cluster/ClusterCli.java
@@ -1,0 +1,175 @@
+package com.wepay.waltz.tools.cluster;
+
+import com.wepay.riff.network.ClientSSL;
+import com.wepay.waltz.client.Transaction;
+import com.wepay.waltz.client.WaltzClient;
+import com.wepay.waltz.client.WaltzClientCallbacks;
+import com.wepay.waltz.client.WaltzClientConfig;
+import com.wepay.waltz.client.internal.InternalRpcClient;
+import com.wepay.waltz.common.util.Cli;
+import com.wepay.waltz.common.util.SubcommandCli;
+import com.wepay.waltz.exception.SubCommandFailedException;
+import com.wepay.waltz.tools.CliConfig;
+import com.wepay.zktools.clustermgr.ClusterManager;
+import com.wepay.zktools.clustermgr.Endpoint;
+import com.wepay.zktools.clustermgr.internal.ClusterManagerImpl;
+import com.wepay.zktools.clustermgr.internal.DynamicPartitionAssignmentPolicy;
+import com.wepay.zktools.clustermgr.internal.PartitionAssignmentPolicy;
+import com.wepay.zktools.clustermgr.internal.ServerDescriptor;
+import com.wepay.zktools.zookeeper.ZNode;
+import com.wepay.zktools.zookeeper.ZooKeeperClient;
+import com.wepay.zktools.zookeeper.internal.ZooKeeperClientImpl;
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.Option;
+import org.apache.commons.cli.Options;
+import org.yaml.snakeyaml.Yaml;
+
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+
+/**
+ * ClusterCli is a tool for interacting with the Waltz Cluster.
+ */
+public final class ClusterCli extends SubcommandCli {
+
+    private ClusterCli(String[] args,  boolean useByTest) {
+        super(args, useByTest, Arrays.asList(
+            new Subcommand(CheckConnectivity.NAME, CheckConnectivity.DESCRIPTION, CheckConnectivity::new)
+        ));
+    }
+
+    /**
+     * The {@code CheckConnectivity} command checks the connectivity
+     * 1. of the server nodes, and
+     * 2. of the storage nodes from each server node
+     * within the Waltz Cluster.
+     */
+    private static final class CheckConnectivity extends Cli {
+        private static final String NAME = "check-connectivity";
+        private static final String DESCRIPTION = "Checks the connectivity of the cluster.";
+
+        private final PartitionAssignmentPolicy partitionAssignmentPolicy = new DynamicPartitionAssignmentPolicy();
+
+        private CheckConnectivity(String[] args) {
+            super(args);
+        }
+
+        @Override
+        protected void configureOptions(Options options) {
+            Option cliCfgOption = Option.builder("c")
+                .longOpt("cli-config-path")
+                .desc("Specify the cli config file path required for zooKeeper connection string, zooKeeper root path and SSL config")
+                .hasArg()
+                .build();
+            cliCfgOption.setRequired(true);
+
+            options.addOption(cliCfgOption);
+        }
+
+        @Override
+        protected void processCmd(CommandLine cmd) throws SubCommandFailedException {
+            ZooKeeperClient zkClient = null;
+            try {
+                String cliConfigPath = cmd.getOptionValue("cli-config-path");
+                CliConfig cliConfig = CliConfig.parseCliConfigFile(cliConfigPath);
+                String zookeeperHostPorts = (String) cliConfig.get(CliConfig.ZOOKEEPER_CONNECT_STRING);
+                String zkRoot = (String) cliConfig.get(CliConfig.CLUSTER_ROOT);
+                int zkSessionTimeout = (int) cliConfig.get(CliConfig.ZOOKEEPER_SESSION_TIMEOUT);
+
+                zkClient = new ZooKeeperClientImpl(zookeeperHostPorts, zkSessionTimeout);
+                ZNode root = new ZNode(zkRoot);
+
+                Set<Endpoint> serverEndpoints = getListOfServerEndpoints(root, zkClient);
+                WaltzClientConfig waltzClientConfig = getWaltzClientConfig(cliConfigPath);
+                Map<Endpoint, Map<String, Boolean>> connectivityStatusMap =
+                    checkServerConnections(waltzClientConfig,
+                        serverEndpoints);
+
+                for (Map.Entry<Endpoint, Map<String, Boolean>> connectivityStatus : connectivityStatusMap.entrySet()) {
+                    System.out.println("Connectivity status of " + connectivityStatus.getKey() + " is: "
+                        + ((connectivityStatus.getValue() == null) ? "UNREACHABLE"
+                        : connectivityStatus.getValue().toString()));
+                }
+            } catch (Exception e) {
+                throw new SubCommandFailedException(String.format("Failed to check topology of the cluster. %n%s",
+                    e.getMessage()));
+            } finally {
+                if (zkClient != null) {
+                    zkClient.close();
+                }
+            }
+        }
+
+        private Set<Endpoint> getListOfServerEndpoints(ZNode root, ZooKeeperClient zkClient) throws Exception {
+            ClusterManager clusterManager = new ClusterManagerImpl(zkClient, root, partitionAssignmentPolicy);
+            Set<ServerDescriptor> serverDescriptors = clusterManager.serverDescriptors();
+            Set<Endpoint> serverEndpoints = new HashSet<>();
+
+            for (ServerDescriptor serverDescriptor: serverDescriptors) {
+                serverEndpoints.add(serverDescriptor.endpoint);
+            }
+            return serverEndpoints;
+        }
+
+        private Map<Endpoint, Map<String, Boolean>> checkServerConnections(WaltzClientConfig config,
+                                                                           Set<Endpoint> serverEndpoints) throws Exception {
+            DummyTxnCallbacks callbacks = new DummyTxnCallbacks();
+            InternalRpcClient rpcClient = new InternalRpcClient(ClientSSL.createContext(config.getSSLConfig()),
+                WaltzClientConfig.DEFAULT_MAX_CONCURRENT_TRANSACTIONS, callbacks);
+
+            return rpcClient.checkServerConnections(serverEndpoints).get();
+        }
+
+        @Override
+        protected String getUsage() {
+            return buildUsage(NAME, DESCRIPTION, getOptions());
+        }
+    }
+
+    /**
+     * Return an object of {@code WaltzClientConfig} built from configuration file.
+     * @param configFilePath the path to configuration file
+     * @return WaltzClientConfig
+     * @throws IOException
+     */
+    private static WaltzClientConfig getWaltzClientConfig(String configFilePath) throws IOException {
+        Yaml yaml = new Yaml();
+        try (FileInputStream in = new FileInputStream(configFilePath)) {
+            Map<Object, Object> props = yaml.load(in);
+            props.put(WaltzClientConfig.AUTO_MOUNT, false);
+            return new WaltzClientConfig(props);
+        }
+    }
+
+    /**
+     * A transaction callback to help construct {@link WaltzClient}. It is dummy because
+     * it is not suppose to receive any callbacks.
+     */
+    private static final class DummyTxnCallbacks implements WaltzClientCallbacks {
+
+        @Override
+        public long getClientHighWaterMark(int partitionId) {
+            return -1L;
+        }
+
+        @Override
+        public void applyTransaction(Transaction transaction) {
+        }
+
+        @Override
+        public void uncaughtException(int partitionId, long transactionId, Throwable exception) {
+        }
+    }
+
+    public static void testMain(String[] args) {
+        new ClusterCli(args, true).processCmd();
+    }
+
+    public static void main(String[] args) {
+        new ClusterCli(args, false).processCmd();
+    }
+}
